Moved several methods from TimeSeriesForest to base class
Several classifier generic methods were moved to the base class for less work when using inheritance.
Constructor keyword "label" has been corrected to "feature". The same goes for TimeSeriesForest.selected_label.
