+++ classifier/timeseries/time_series_forest.py
@@ -1,10 +1,8 @@
 from typing import Tuple
 
 import numpy as np
-import matplotlib.pyplot as plt
 import pandas as pd
 from sklearn.model_selection import train_test_split
-from sklearn.metrics import confusion_matrix, classification_report, ConfusionMatrixDisplay
 from sktime.classification.interval_based import TimeSeriesForestClassifier
 
 from objects.training_set import TrainingSet, Patient
@@ -13,16 +11,16 @@ from classifier.timeseries.time_series_classifier import TimeSeriesClassifier
 
 class TimeSeriesForest(TimeSeriesClassifier):
 
-    def __init__(self, data_set: TrainingSet, train_fraction: float = 0.8, label: str = "HR"):
+    def __init__(self, data_set: TrainingSet, train_fraction: float = 0.8, feature: str = "HR"):
         super().__init__(data_set=data_set, train_fraction=train_fraction)
-        self.selected_label = label
+        self.selected_feature = feature
         self.model = TimeSeriesForestClassifier(random_state=1337, n_jobs=-1)
         self.train_data = None
         self.test_data = None
 
     def transform_data_set(self, data_set: TrainingSet, label_set: pd.DataFrame) -> tuple:
         transformed_data_set = data_set.get_timeseries_df(use_interpolation=True, fix_missing_values=True,
-                                                          limit_to_features=[self.selected_label])
+                                                          limit_to_features=[self.selected_feature])
         transformed_labels = []
         for patient_id, label in transformed_data_set.index:
             transformed_labels.append(label_set[patient_id][0])
@@ -53,33 +51,6 @@ class TimeSeriesForest(TimeSeriesClassifier):
         y_prediction = self.model.predict(x_data)
         return y_prediction
 
-    def test(self, x_data, y_data):
-        return self.get_confusion_matrix(y_data, self.predict(x_data))
 
-    def get_confusion_matrix(self, y_data, y_predicted) -> np.ndarray:
-        cm = confusion_matrix(y_data, y_predicted)
-        return cm
-
-    def get_classification_report(self, x_data, y_data):
-        report = classification_report(y_data, self.predict(x_data))
-        return report
-
-    def display_confusion_matrix(self, test_set=None, plotting: bool = False):
-        if test_set is None:
-            test_set = self.test_data
-        x_test, y_test = test_set
-
-        # classification_report
-        report = self.get_classification_report(x_test, y_test)
-        print(report)
-        # confusion matrix plot
-        if plotting:
-            cm: np.ndarray = self.test(x_test, y_test)
-            disp = ConfusionMatrixDisplay(confusion_matrix=cm, display_labels=["No Sepsis", "Sepsis"])
-            disp.plot()
-            # funktioniert leider nicht mit title
-            # temp_text_obj = Text(x=100, y=50, text=f"Confusion Matrix for version: {version}")
-            # disp.ax_.title = temp_text_obj
-            plt.show()
